start of engine 2.0
This was a complete rewrite that took around 2 weeks to complete. 

I chose to rewrite the engine because the way I was doing it before was too messy and didn't use as many bitboards as it should have. The new way is significantly faster for move generation and fits a more standard chess engine design, drawing inspiration from the design and architecture of the most powerful engine, stockfish.